CuraTrack System/requirements.txt
Flask
Flask-Login
mysql-connector-python